fix: compilation errors and structural connections
